Create and separate Controller to a new package
To fix Junit testing error
